% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TUSCAN_utilities.R
\name{createTUSCANObject}
\alias{createTUSCANObject}
\title{Create a TUSCAN object}
\usage{
createTUSCANObject(counts, location, project = "TUSCAN", img, r = 49)
}
\arguments{
\item{counts}{Raw spatial transcriptomics gene expression count matrix, the dimension is \eqn{m \times n}, where \eqn{m} is the number of genes and \eqn{n} is the number of spots.}

\item{location}{Dataframe containing the spatial coordinates of all spots, where each row represents a spot and two columns represent the \eqn{x} and \eqn{y} coordinates of the spatial location, respectively.}

\item{project}{Name of the project.}

\item{img}{Histology image input.}

\item{r}{Image smoothing parameter: draw a square with a side length of (\eqn{2r+1}) centered on each spot. Apply the average color within the square as the new RGB values. The default value is 49.}
}
\value{
Return a preliminary TUSCAN object.
}
\description{
Create a TUSCAN object.
}
