% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindNormalCluster.R
\name{FindNormalCluster}
\alias{FindNormalCluster}
\title{Find a subset of normal spots}
\usage{
FindNormalCluster(object, w = 0.5)
}
\arguments{
\item{object}{TUSCAN object.}

\item{w}{Weight of image. Default value is 0.5, i.e., gene expression and image are equally weighted.}
}
\value{
Print the name of the selected normal cluster.
}
\description{
This function helps determine the cluster with the highest potential purity of normal spots through an evaluation score \eqn{s} that combines both gene expression and histology image information.
The evaluation score \eqn{s} of each cluster is calculated as \eqn{s=w \times g - (1-w) \times \sigma}, where \eqn{g} is the median gray value of all spots within that cluster extracted from the gray-scaled histology image, \eqn{\sigma} is standard deviation of all gene expression levels within that cluster, and \eqn{w} is the weight of image. Both \eqn{g} and \eqn{\sigma} are standardized.
The cluster with the largest \eqn{s} is selected as the normal reference.
}
