% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatial.R
\name{plotSpatial}
\alias{plotSpatial}
\title{Spatial plotting functions}
\usage{
plotSpatial(obj, class = c("cluster", "tumor"), color = NULL, size)
}
\arguments{
\item{obj}{TUSCAN object.}

\item{color}{Optional color palette can be customized by user.}

\item{size}{Point size.}

\item{cluster}{Options: "cluster","tumor".}
}
\value{
Return a spatial plot
}
\description{
Spatial plotting functions
}
