% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreCluster.R
\name{PreCluster}
\alias{PreCluster}
\title{Preliminary clustering integrating spatial gene expression data and histology image}
\usage{
PreCluster(object, res = NULL, method = c("louvain", "km"), k = 5, seed = NULL)
}
\arguments{
\item{object}{TUSCAN object.}

\item{res}{Optional resolution parameter in louvain clustering that allows the user to adjust the number of clusters. Larger values typically yield more clusters.}

\item{method}{Clustering method. Options: c('louvain',"km")}

\item{k}{Number of clusters from K-Means.}

\item{seed}{Optional numerical value. Set a random seed for reproducible results.}
}
\value{
Return a processed TUSCAN object adding preliminary clustering results.
}
\description{
This function performs a preliminary clustering of all spots based on spatial gene expression matrix and histology image.
The aim of this step is to select a subset of normal spots as the reference to infer copy number profile later.
We recommend using a larger number of clusters to make the selected reference group is purer.
We recommend that the reference group contains at least 10 spots to reduce potential bias.
}
