% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreCluster.R
\name{reduceDim}
\alias{reduceDim}
\title{Perform dimension reduction by autoencoder}
\usage{
reduceDim(
  object,
  nhvg = 2000,
  d = 20,
  epochs = 50,
  batch_size = 64,
  seed = NULL
)
}
\arguments{
\item{object}{TUSCAN object.}

\item{nhvg}{Number of highly variable genes to run autoencoder.}

\item{d}{Number of dimensions compressed by autoencoder.}

\item{epochs}{Number of training iterations.}

\item{batch_size}{Number of training samples utilized during one iteration.}

\item{seed}{Optional numerical value. Set a random seed for reproducible results.}
}
\value{
Return a processed TUSCAN object adding low-dimensional components.
}
\description{
Perform dimension reduction by autoencoder.
}
