% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TumorCluster.R
\name{TumorCluster}
\alias{TumorCluster}
\title{Predict tumor spots}
\usage{
TumorCluster(
  object,
  seed = NULL,
  maxK = 3,
  reps = 100,
  pItem = 0.8,
  pFeature = 1,
  title = NULL,
  clusterAlg = c("km", "hc", "pam"),
  distance = c("euclidean", "pearson", "spearman", "maximum", "canberra", "minkowski")
)
}
\arguments{
\item{object}{TUSCAN object.}

\item{seed}{Optional numerical value. Set a random seed for reproducible results.}

\item{maxK}{Maximum cluster number to evaluate.}

\item{reps}{Number of subsamples.}

\item{pItem}{Proportion of items to sample.}

\item{pFeature}{Proportion of features to sample.}

\item{title}{File directory for storing consensus clustering figures.}

\item{clusterAlg}{Cluster algorithm. (default: 'km'), options('km','hc','pam'). 'km': k-means, 'hc': hierarchical (hclust), 'pam': paritioning around medoids.}

\item{distance}{Distance function. (default: 'euclidean'), options('euclidean','pearson','spearman','maximum','canberra','minkowski').}
}
\value{
Return a a processed TUSCAN object adding tumor annotations.
}
\description{
This function implements the automatic tumor segmentation and classification.
}
