% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TUSCAN_utilities.R
\docType{class}
\name{TUSCAN-class}
\alias{TUSCAN-class}
\title{The TUSCAN Class}
\description{
Slots in the TUSCAN object include:
}
\section{Slots}{

\describe{
\item{\code{counts.data}}{The raw expression count matrix. Rows are genes, columns are spots.}

\item{\code{cnv.data}}{The copy number matrix inferred from the raw count matrix. Rows are genes, columns are spots.}

\item{\code{project}}{Name of the project.}

\item{\code{location}}{The spatial coordinates of the spots, with two columns representing the \eqn{x} and \eqn{y} coordinates of the spatial location.}

\item{\code{image}}{The histology image information, including the pixel location of each spot on the original histology image, and color information.}

\item{\code{autoencoder}}{The outputs from autoencoder.}

\item{\code{pre_cluster}}{The preliminary clustering results that are used to select a reference cluster later.}

\item{\code{tumor_cluster}}{Tumor prediction, indicating each spot as 'tumor' or 'normal'.}

\item{\code{annotation}}{Annotations of all spots, with two columns representing the spot barcode and the spot clustering name. If the user do not provide an annotation file, the \code{pre_cluster} result will be used as the annotations.}

\item{\code{gene_order}}{The gene order file, contains chromosome, start, and stop position for each gene.}

\item{\code{reference_grouped_spot_indices}}{The indices of the reference spots.}

\item{\code{observation_grouped_spot_indices}}{The indices of the other spots.}
}}

