% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TUSCAN.R
\docType{package}
\name{TUSCAN-package}
\alias{TUSCAN}
\alias{TUSCAN-package}
\title{The 'TUSCAN' package}
\description{
Roxygen2 allows you to write documentation in comment blocks co-located
with code.
}
\details{
The only function you're likely to need from \pkg{roxygen2} is
\code{\link{roxygenize}}. Otherwise refer to the vignettes to see
how to format the documentation.
}
\author{
\strong{Maintainer}: Chenxuan Zang \email{czang@mdanderson.org}

Authors:
\itemize{
  \item Peng Wei \email{pwei2@mdanderson.org}
  \item Ziyi Li \email{zli16@mdanderson.org}
  \item Charles C. Guo \email{ccguo@mdanderson.org}
}

}
